도서 마이 페이지 기능 구현완료 (대여목록, 위리리스트, 리뷰관련) 동은

+++ codeIt/src/main/java/com/codeit/mini/controller/book/RestAPIController.java
@@ -1,7 +1,6 @@
 package com.codeit.mini.controller.book;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -12,20 +11,27 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.codeit.mini.dto.book.BookDTO;
+import com.codeit.mini.dto.book.ReviewDTO;
+import com.codeit.mini.dto.member.MemberDTO;
 import com.codeit.mini.entity.book.BookEntity;
 import com.codeit.mini.entity.book.RentEntity;
+import com.codeit.mini.entity.book.ReviewEntity;
 import com.codeit.mini.entity.book.WishEntity;
 import com.codeit.mini.repository.book.IBookRepository;
 import com.codeit.mini.service.book.IBookSearchService;
 import com.codeit.mini.service.book.IBookService;
 import com.codeit.mini.service.book.IRentService;
+import com.codeit.mini.service.book.IReviewService;
 import com.codeit.mini.service.book.IWishService;
 
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
@@ -41,6 +47,7 @@ public class RestAPIController {
 	private final IBookService bookService;
 	private final IRentService rentService;
 	private final IWishService wishService;
+	private final IReviewService reviewService;
 	
 	// 도서 검색
 	@GetMapping("/search")
@@ -59,8 +66,15 @@ public class RestAPIController {
 	
 	// 도서 대여
 	@PostMapping("/rent")
-	public ResponseEntity<RentEntity> rentBook(@RequestParam("bookId") Long bookId,
-									  		   @RequestParam("memberId") Long memberId) {
+	public ResponseEntity<?> rentBook(@RequestParam("bookId") Long bookId, HttpSession session) {
+		MemberDTO member = (MemberDTO)session.getAttribute("member");
+		
+		if (member == null) {
+			return new ResponseEntity<>("로그인 후 이용해주세요.", HttpStatus.UNAUTHORIZED);
+		}
+		
+		Long memberId = member.getMemberId();
+		
 		try {
 			RentEntity rentEntity = rentService.rentBook(bookId, memberId);
 			return new ResponseEntity<>(rentEntity, HttpStatus.OK);
@@ -71,8 +85,14 @@ public class RestAPIController {
 	
 	// 도서 찜추가
 	@PostMapping("/addWish")
-	public ResponseEntity<?> addWish(@RequestParam("bookId") Long bookId,
-									 @RequestParam("memberId") Long memberId) {
+	public ResponseEntity<?> addWish(@RequestParam("bookId") Long bookId, HttpSession session) {
+		MemberDTO member = (MemberDTO) session.getAttribute("member");
+		
+		if (member == null) {
+			return new ResponseEntity<>("로그인 후 이용해주세요.", HttpStatus.UNAUTHORIZED);
+		}
+		
+		Long memberId = member.getMemberId();
 		try {
 			WishEntity wishEntity = wishService.addWished(bookId, memberId);
 			return new ResponseEntity<>(wishEntity, HttpStatus.OK);
@@ -91,8 +111,15 @@ public class RestAPIController {
 	
 	// 도서 찜 삭제
 	@DeleteMapping("/removeWish")
-	public ResponseEntity<String> removeWish(@RequestParam("bookId") Long bookId,
-			  								 @RequestParam("memberId") Long memberId) {
+	public ResponseEntity<?> removeWish(@RequestParam("bookId") Long bookId, HttpSession session) {
+		MemberDTO member = (MemberDTO)session.getAttribute("member");
+		
+		if (member == null) {
+			return new ResponseEntity<>("로그인 후 이용해주세요.", HttpStatus.UNAUTHORIZED);
+		}
+		
+		Long memberId =  member.getMemberId();
+		
 		try {
 			boolean removed = wishService.removeWished(bookId, memberId);
 			if (removed) {
@@ -111,17 +138,98 @@ public class RestAPIController {
 		}
 	}
 	
-	// 도서 필요 포인트 출력
-//	@GetMapping()
-//	public ResponseEntity<Map<String, Integer>> getBookCost(@PathVariable Long bookId) throws Exception {
-//		
-//		BookEntity bookEntity = bookRepository.findById(bookId)
-//							.orElseThrow(()-> new IllegalAccessException("해당 도서가 없습니다."));
-//		
-//		Map<String, Integer> rs = new HashMap<>();
-//		rs.put("pointCost", bookEntity.getRentPoint());
-//		
-//		return new ResponseEntity<>(rs, HttpStatus.OK);
-//	}
+	// 리뷰 등록
+	@PostMapping("/regReview")
+	public ResponseEntity<?> regReview(@RequestBody ReviewDTO reviewDTO, HttpSession session) {
+		
+		MemberDTO member = (MemberDTO) session.getAttribute("member");
+		
+		if (member == null) {
+			return new ResponseEntity<>("로그인 후 이용해주세요.", HttpStatus.UNAUTHORIZED);
+		}
+		
+		try {
+			reviewDTO.setMemberId(member.getMemberId());
+			ReviewEntity reviewEntity = reviewService.regReview(reviewDTO);
+			return new ResponseEntity<>(reviewEntity, HttpStatus.OK);
+			
+		} catch (IllegalStateException e) {
+			return new ResponseEntity<>("리뷰가 있습니다.", HttpStatus.CONFLICT);
+		} catch (IllegalAccessException e) {
+			return new ResponseEntity<>("NOT RENT", HttpStatus.BAD_REQUEST);
+		} catch (Exception e) {
+			return new ResponseEntity<>("SERVER ERROR", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+	
+	// 리뷰 삭제
+	@DeleteMapping("/removeReview/{reviewId}")
+	public ResponseEntity<?> removeReview(@PathVariable("reviewId") Long reviewId, HttpSession session) {
+		MemberDTO member = (MemberDTO) session.getAttribute("member");
+		if (member == null) {
+			return new ResponseEntity<>("로그인 후 이용해주세요.", HttpStatus.UNAUTHORIZED);
+		}
+		
+		try {
+			Long memberId = member.getMemberId();
+			
+			ReviewDTO reviewDTO = reviewService.findById(reviewId);
+			
+			if (reviewDTO == null) {
+				return new ResponseEntity<>("NOT REVIEW", HttpStatus.NOT_FOUND);
+			}
+			
+			if (!reviewDTO.getMemberId().equals(memberId)) {
+				return new ResponseEntity<>("NOT MEMBER", HttpStatus.FORBIDDEN);
+			}
+			
+			reviewService.removeReview(reviewId);
+			
+			return new ResponseEntity<>("REMOVE REVIEW", HttpStatus.OK);
+		} catch (Exception e) {
+			e.printStackTrace();
+			return new ResponseEntity<>("SERVER ERROR", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+		
+	// 리뷰 수정
+	@PutMapping("/modifyReview/{reviewId}")
+	public ResponseEntity<?> modifyReview(@PathVariable("reviewId") Long reviewId,
+										  @RequestBody ReviewDTO reviewDTO,
+										  HttpSession session) {
+		try {
+			MemberDTO member = (MemberDTO) session.getAttribute("member");
+			if (member == null) {
+				return new ResponseEntity<>("로그인 후 이용해주세요.", HttpStatus.UNAUTHORIZED);
+			}
+			
+			reviewDTO.setReviewId(reviewId);
+			reviewDTO.setMemberId(member.getMemberId());
+			
+			reviewService.modifyReview(reviewDTO);
+			return new ResponseEntity<>("MODIFY REVIEW", HttpStatus.OK);
+		} catch (Exception e) {
+			log.error("SERVER ERROR", e.getMessage());
+			return new ResponseEntity<>("SERVER ERROR", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
+	
+	// 회원 리뷰 리스트 출력
+	@GetMapping("/myReviewList/{bookId}")
+	public ResponseEntity<?> getMyReviewList(@PathVariable("bookId") Long bookId, HttpSession session) {
+		MemberDTO member =(MemberDTO) session.getAttribute("member");
+		
+		if (member == null) {
+			return new ResponseEntity<>("로그인 후 이용해주세요.", HttpStatus.UNAUTHORIZED);
+		}
+		try {
+			Long memberId = member.getMemberId();
+			List<ReviewDTO> review = reviewService.findReviewByMemberId(bookId, memberId);
+			return new ResponseEntity<>(review, HttpStatus.OK);
+		} catch (Exception e) {
+			e.printStackTrace();
+			return new ResponseEntity<>("리뷰 조회 중 에러", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+	}
 	
 }
